voice_assistant_requirements.txt
discord.py>=2.3.2
SpeechRecognition>=3.10.0
openai-whisper>=20231117
azure-cognitiveservices-speech>=1.31.0
pydub>=0.25.1
yt-dlp>=2023.11.14
PyNaCl>=1.5.0
numpy>=1.24.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0 